Scratch CI using GitHub actions: lint & test
Sadly we can't use official gitlab-ci action cos it's based on linux container,
but we need mingw-64 and Windows headers even for linter. Moreover golangci-lint
works so bad on Windows so it seems currently easier to use cross-build container.

Tests on the other hand runs on the honest Windows VM.